+++ shopos-core/src/Core/Plugin.php
@@ -14,7 +14,7 @@ defined( 'ABSPATH' ) || exit;
  */
 final class Plugin {
 
-	const VERSION = '1.50.0';
+	const VERSION = '1.51.0';
 
 	/**
 	 * Singleton instance.